+++ tests/__helpers__/data.ts
@@ -1,32 +1,20 @@
 import {
-  CreateMode,
-  ReadMode,
-  UpdateMode,
-  DeleteMode,
-  GroupSyncMode,
-  IndexFieldsType,
-  CreationTransitionType,
-  ConfigurationType,
   Condition,
-  Transition,
-  CreationTransitionAction,
-  CreationTransitionAfterAction,
   CreationTransition,
+  Index,
   Schema,
-  InitiatorHasRelationToUserInDataConditionType,
-  InitiatorHasRelationToUserInDataConditionRelation,
-  InitiatorHasRelationToGroupInDataConditionType,
-  InitiatorHasRelationToGroupInDataConditionRelation,
+  SchemaInput,
+  TransitionInput,
 } from '../../src/services/data/types';
 
-export const newSchemaInput = {
+export const newSchemaInput: SchemaInput = {
   name: 'Fibricheck measurement',
   description: 'The schema for holding FibriCheck measurements',
-  createMode: CreateMode.DEFAULT,
-  readMode: ReadMode.ALL_USERS,
-  updateMode: UpdateMode.DEFAULT,
-  deleteMode: DeleteMode.PERMISSION_REQUIRED,
-  groupSyncMode: GroupSyncMode.DISABLED,
+  createMode: 'default',
+  readMode: 'allUsers',
+  updateMode: 'default',
+  deleteMode: 'permissionRequired',
+  groupSyncMode: 'disabled',
   defaultLimit: 5,
   maximumLimit: 5,
 };
@@ -65,105 +53,100 @@ export const newSchemaCreated: Partial<Schema> = {
   },
   creationTransition: {
     toStatus: 'start',
-    type: CreationTransitionType.MANUAL,
+    type: 'manual',
     conditions: [
       {
-        type: ConfigurationType.INPUT,
+        type: 'input',
         configuration: {
-          type: ConfigurationType.NUMBER,
+          type: 'number',
           minimum: -180,
           maximum: 180,
         },
       },
       {
-        type: ConfigurationType.DOCUMENT,
+        type: 'document',
         configuration: {
-          type: ConfigurationType.NUMBER,
+          type: 'number',
           minimum: -180,
           maximum: 180,
         },
       },
       {
-        type:
-          InitiatorHasRelationToUserInDataConditionType.INITIATOR_HAS_RELATION_TO_USER_IN_DATA,
+        type: 'initiatorHasRelationToUserInData',
         userIdField: '5e9fff9d90135a2a9a718e2f',
-        relation:
-          InitiatorHasRelationToUserInDataConditionRelation.IS_STAFF_OF_TARGET_PATIENT,
+        relation: 'isStaffOfTargetPatient',
       },
       {
-        type:
-          InitiatorHasRelationToGroupInDataConditionType.INITIATOR_HAS_RELATION_TO_GROUP_IN_DATA,
+        type: 'initiatorHasRelationToGroupInData',
         groupIdField: '5e9fff9d90135a2a9a718e2f',
-        relation: InitiatorHasRelationToGroupInDataConditionRelation.STAFF,
+        relation: 'staff',
         requiredPermission: 'MY_PERMISSION',
       },
     ],
     actions: [
       {
-        type: CreationTransitionAction.ALGORITHM,
+        type: 'task',
+        functioName: 'test',
+        data: {},
       },
     ],
     afterActions: [
       {
-        type: CreationTransitionAfterAction.NOTIFY_ALGO_QUEUE_MANAGER,
+        type: 'notifyAlgoQueueManager',
+        id: 'id',
+        version: 'version',
       },
     ],
   },
   transitions: [
     {
+      id: '1',
       toStatus: 'start',
-      type: CreationTransitionType.MANUAL,
+      type: 'manual',
       conditions: [
         {
-          type: ConfigurationType.INPUT,
+          type: 'input',
           configuration: {
-            type: ConfigurationType.NUMBER,
+            type: 'number',
             minimum: -180,
             maximum: 180,
           },
         },
         {
-          type: ConfigurationType.DOCUMENT,
+          type: 'document',
           configuration: {
-            type: ConfigurationType.NUMBER,
+            type: 'number',
             minimum: -180,
             maximum: 180,
           },
         },
         {
-          type:
-            InitiatorHasRelationToUserInDataConditionType.INITIATOR_HAS_RELATION_TO_USER_IN_DATA,
+          type: 'initiatorHasRelationToUserInData',
           userIdField: '5e9fff9d90135a2a9a718e2f',
-          relation:
-            InitiatorHasRelationToUserInDataConditionRelation.IS_STAFF_OF_TARGET_PATIENT,
+          relation: 'isStaffOfTargetPatient',
         },
         {
-          type:
-            InitiatorHasRelationToGroupInDataConditionType.INITIATOR_HAS_RELATION_TO_GROUP_IN_DATA,
+          type: 'initiatorHasRelationToGroupInData',
           groupIdField: '5e9fff9d90135a2a9a718e2f',
-          relation: InitiatorHasRelationToGroupInDataConditionRelation.STAFF,
+          relation: 'staff',
           requiredPermission: 'MY_PERMISSION',
         },
       ],
       actions: [
         {
-          type: CreationTransitionAction.ALGORITHM,
+          type: 'linkCreator',
         },
       ],
-      afterActions: [
-        {
-          type: CreationTransitionAfterAction.NOTIFY_ALGO_QUEUE_MANAGER,
-        },
-      ],
+      afterActions: [],
       name: 'move',
       fromStatuses: ['start'],
     },
   ],
-  createMode: CreateMode.DEFAULT,
-  readMode: ReadMode.ALL_USERS,
-  updateMode: UpdateMode.DEFAULT,
-  deleteMode: DeleteMode.PERMISSION_REQUIRED,
-  groupSyncMode: GroupSyncMode.DISABLED,
+  createMode: 'default',
+  readMode: 'allUsers',
+  updateMode: 'default',
+  deleteMode: 'permissionRequired',
+  groupSyncMode: 'disabled',
   defaultLimit: 5,
   maximumLimit: 5,
   updateTimestamp: new Date('2021-04-20T07:34:08.358Z'),
@@ -330,11 +313,11 @@ export const newIndexCreated = {
   system: true,
 };
 
-export const newIndexInput = {
+export const newIndexInput: Index = {
   fields: [
     {
       name: 'PropertyNameToIndex',
-      type: IndexFieldsType.ASC,
+      type: 'asc',
     },
   ],
   options: {
@@ -442,18 +425,18 @@ export const lockedDocumentsListResponse = {
 };
 
 const inputCondition: Condition = {
-  type: ConfigurationType.INPUT,
+  type: 'input',
   configuration: {
-    type: ConfigurationType.NUMBER,
+    type: 'number',
     minimum: -180,
     maximum: 180,
   },
 };
 
 const documentCondition: Condition = {
-  type: ConfigurationType.DOCUMENT,
+  type: 'document',
   configuration: {
-    type: ConfigurationType.NUMBER,
+    type: 'number',
     minimum: -180,
     maximum: 180,
   },
@@ -461,22 +444,24 @@ const documentCondition: Condition = {
 
 export const transitionInput: CreationTransition = {
   toStatus: 'start',
-  type: CreationTransitionType.MANUAL,
+  type: 'manual',
   conditions: [inputCondition, documentCondition],
 };
 
-export const newTransition: Transition = {
+export const newTransition: TransitionInput = {
   toStatus: 'start',
-  type: CreationTransitionType.MANUAL,
+  type: 'manual',
   conditions: [inputCondition, documentCondition],
   actions: [
     {
-      type: CreationTransitionAction.ALGORITHM,
+      type: 'linkCreator',
     },
   ],
   afterActions: [
     {
-      type: CreationTransitionAfterAction.NOTIFY_ALGO_QUEUE_MANAGER,
+      type: 'notifyAlgoQueueManager',
+      id: 'id',
+      version: 'version',
     },
   ],
   name: 'move',