+++ CHANGELOG.md
@@ -34,7 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -    };
 -  };
 - }
-- const document = await sdk.data.findDocuments<CustomDocument>();
+- const document = await sdk.data.documents.find<CustomDocument>();
 
 + interface MyData {
 +   data: {
@@ -45,12 +45,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 +     };
 +   };
 + }
-+ const document = await sdk.data.findDocuments<MyData>();
++ const document = await sdk.data.documents.find<MyData>();
 ```
 
 ### Added
 
-- `sdk.files.createRaw` where you can pass in your FormData directly
+- `sdk.files.createFromText` where you can pass in your text directly.
+
+```ts
+await sdk.files.createFromText('this-is-a-string');
+```
 
 - Added `transitionsByName` getter to easily get the transition you need.