+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Localizations Service
 - Added Profiles Service
 - Added Notifications Service
+- Added optional `skipTokenCheck` parameter to Oauth1 authentication flow with token/tokenSecret.
 
 ## [3.1.0]